docbook: fix signedness
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@35463 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ plugins/docbook/xp/ie_imp_DocBook.h
@@ -82,7 +82,7 @@ protected:
 	int m_iFootnotes;
 	int m_iImages;
 	int m_iSectionDepth;
-	int m_iTitleDepth;
+	size_t m_iTitleDepth;
 	UT_sint32 m_iNoteID;
 	std::vector<fl_AutoNumPtr> m_utvTitles;
 	bool m_bMustAddTitle;
